Type mismatch rectified

+++ tests/const_folding.go
@@ -1,7 +1,7 @@
 package cf
 
 func main() {
-	var a int = 10 + 20 / 2;
+	var a float64 = 10 + 20/2;
 	var b int = (13 - 12) * 5;
 	var c bool = 2 >= 1
 	var d bool = 14 == (15-1)
